+++ src/main/java/com/sohwakmo/hr/repository/MeetingRoomRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface MeetingRoomRepository extends JpaRepository<MeetingRoom, Integer> {
     List<MeetingRoom> findAll();
 
-    List<MeetingRoom> findByAttendeeOrReservationNo(Integer attendee, Integer reservationNo);
+    List<MeetingRoom> findByAttendeeOrReservationNo(Long attendee, Long reservationNo);
 }
